Fix double HTTP host header
HTTPConnection.putrequest has a skip_host flag that is false by default. When false, the host is automatically added to the HTTP header. The host is also added manually so there is currently two occurrences in the header. This may cause issues depending on how the receiver handles recurring headers.

This change simply sets the skip_host flag to true.